fix(db): Resolve Room migration crash on startup
Resolves a fatal crash on startup caused by a missing Room database migration from version 24 to 25.

The database initialization in the Hilt AppModule was not configured with the necessary migrations. This change moves the complete database builder logic into AppModule.kt, ensuring that the singleton database instance is always created with the correct migration paths and other required configurations.

The redundant newInstance method in MusicDatabase.kt has been removed to prevent future misuse and centralize database creation.
